+++ src/test/java/io/nagurea/smsupsdk/webhooks/delete/DeleteWebhookServiceIntTest.java
@@ -1,0 +1,88 @@
+package io.nagurea.smsupsdk.webhooks.delete;
+
+import io.nagurea.smsupsdk.common.status.ResponseStatus;
+import io.nagurea.smsupsdk.webhooks.delete.response.DeleteWebhookResponse;
+import io.nagurea.smsupsdk.webhooks.delete.response.DeleteWebhookResultResponse;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockserver.integration.ClientAndServer;
+import org.mockserver.model.HttpResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.io.IOException;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockserver.model.HttpRequest.request;
+
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = SpringConfiguration.class)
+class DeleteWebhookServiceIntTest {
+
+    private static final String YOUR_TOKEN = "Your Token";
+    private static final String EXPECTED_TOKEN = "Bearer " + YOUR_TOKEN;
+    private static final String WEBHOOK_ID = "66";
+
+    /**
+     * Useless. Only here to see how services could be used with Spring
+     */
+    @Autowired
+    private DeleteWebhookService deleteWebhookService;
+
+    private static ClientAndServer mockServer;
+
+    @BeforeAll
+    public static void startMockSMSUpServer(){
+        mockServer = ClientAndServer.startClientAndServer("localhost", 4242, 4242);
+        mockServer.when(
+                request()
+                        .withPath("/webhook/" + WEBHOOK_ID)
+                        .withMethod("DELETE")
+                        .withHeader("Authorization", EXPECTED_TOKEN)
+        ).respond(
+                HttpResponse.response()
+                        .withStatusCode(200)
+                        .withBody(
+                                "{\n" +
+                                        "  \"status\": 1,\n" +
+                                        "  \"message\": \"OK\",\n" +
+                                        "  \"deleted_id\": \"66\"\n" +
+                                        "}"
+                        )
+        );
+
+    }
+
+    @AfterAll
+    static void stopMockserver(){
+        mockServer.stop();
+    }
+
+     @Test
+     void deleteList() throws IOException {
+         //given
+         final DeleteWebhookResultResponse expectedResponse = DeleteWebhookResultResponse.builder()
+                 .message(ResponseStatus.OK.getDescription())
+                 .responseStatus(ResponseStatus.OK)
+                 .deletedId("66")
+                 .build();
+         final int expectedStatusCode = 200;
+
+         //when
+         final DeleteWebhookResponse result = deleteWebhookService.deleteWebhook(YOUR_TOKEN, WEBHOOK_ID);
+         final Integer effectiveStatusCode = result.getStatusCode();
+         final DeleteWebhookResultResponse effectiveResponse = result.getEffectiveResponse();
+
+         //then
+         assertEquals(expectedStatusCode, effectiveStatusCode);
+         assertEquals(expectedResponse, effectiveResponse);
+         assertEquals(expectedResponse.getMessage(), effectiveResponse.getMessage());
+         assertEquals(expectedResponse.getStatus(), effectiveResponse.getStatus());
+
+    }
+
+
+}